Create README.md
TeKwonDo Sparring Score Keeper App.

The App allows user to review the scorring rules before the start of the match. The user has an option to enter the competitors' names. The user starts the time. Timer has an option to pause and resume. The score is kept by clicking areas of contact. There are three areas: head, body, and below the belt. Kick to the head - two points. Kick to the body - one point. Kick below the belt is a strike. Clicking on the area of one competitor gives points/strikes to the other competitor. App is locked in landscape mode to ensure consistent experience.